refactor: remove unused layout server file, update homepage title, and add project page components
