--- src/routes/(stuff)/+layout.server.ts
@@ -1,8 +1,0 @@
-import { getUser } from '$lib/server/utils';
-import type { LayoutServerLoad } from './$types';
-
-export const load: LayoutServerLoad = async () => {
-	const user = getUser();
-
-	return { user };
-};
