+++ src/routes/api/delete/+server.ts
