+++ src/routes/home/+page.svelte
@@ -1,0 +1,9 @@
+<script lang="ts">
+	import Toucan from '$lib/assets/toucan.png?enhanced';
+</script>
+
+<h1>Hi</h1>
+<a href="/login" class="font-bold underline">Sign In</a>
+<enhanced:img src={Toucan} alt="Toucan" class="w-100" />
+<p>This is a Toucan. It rhymes with YouKan.</p>
+<p>This homepage is terrible</p>
